backend/requirements.txt
docx==0.2.4
fastapi==0.115.11
PyMuPDF==1.21.0
ImageHash==4.3.2
openai==1.68.2
Pillow==11.1.0
pydantic==2.10.6
PyJWT==2.10.1
python-dotenv==1.0.1
bcrypt==4.0.1
python-docx==1.1.2
qdrant_client==1.13.3
SQLAlchemy==2.0.25
starlette==0.46.1
uvicorn==0.34.0
python-multipart



